Fix connection update errors when already existent
If a connection was not reconciled correctly and had missing identity /
parameter status fields, while a connection was created in the instance,
the reconciling would fail with "already existing" errors. Now, the
reconciliation does not rely on existing status information and infers
necessary parameters. This will also check for already existent
connections in the instance.

If a connection is managed by a CR it will now take control of existing
connections within the instance, even if it was created manually.

Connections can now change parent groups. If a connection changes its
parent and a conflicting connection is already in-place, no
modifications will occur and the reconciliation fails. This corresponds
to the upstream behavior.

Remove unnecessary log of connection resource name and namespace.