Fix connection permission handling
User and group permissions can be added to a connection. If this
connection is in a (nested) connection group, Guacamole requires to have
permissions on all connection groups up the hierarchy. This patch adds
this behavior.

This patch does not handle removing permissions from connection
groups. The permission will be removed from the connection itself, but
not all connection groups up the hierarchy. This is not a problem, as
Guacamole seems not to apply connection group permissions to all
connections within it.

The removal needs to be handled in another patch to avoid cluttering
the home screen with empty connection groups.